requirements.txt
airefinery-sdk==1.21.0
fastapi==0.119.1